Refactoring as Python package structure
