+++ README.md
@@ -1,12 +1,12 @@
 # pylabfront
 
-In this repository you will find code that you can use to analyse data collected through [Labfront](https://labfront.com). Labfront is a platform that can be used to easily collect data from wearable devices. One of the key issues in using Labfront, in comparison to other platforms for physiological data collection and visualization (see: [Empatica Care](https://www.empatica.com/en-eu/care/)) is that it does not offer an appropriate dashboard for the visualization of data, and only shows you if the users have synced their smartwatches with the app. The aim of this repo is to offer a dashboard for:
-- visualization of physiological data at different resolutions (day, week, month, year)
+In this repository you will find code that you can use to analyse data collected through [Labfront](https://labfront.com). Labfront is a platform that can be used to easily collect data from wearable devices. One of the key issues in using Labfront, in comparison to other platforms for physiological data collection and visualization (see: [Empatica Care](https://www.empatica.com/en-eu/care/)) is that it does not offer an appropriate dashboard for the visualization of data, and only shows you if the users have synced their smartwatches with the app. The aim of this repo is to offer a command line interface (CLI) for data extraction and a web-based dashboard for:
+- visualization of physiological data
 - automatic analysis of data
 - extraction of features
 - ...
 
-At the moment, the repo is a playground for understanding how data are saved and organized in Labfront. Hopefully, this will turno into a Python package.
-
 Current features:
 - [x] Extraction of data within a given time range
+- [] Command line interface for extraction of data
+- [] Basic web dashboard for data visualization
